Added changes to WFH

+++ backend/controllers/wfh.js
@@ -3,40 +3,38 @@ import { UserModel } from '../models/userSchema.js'
 import { handleCatch, createNew, updateById, getById, deleteById } from '../utils/common.js'
 
 export const creatingWFH = (req, res, next) => {
-    try {
-        if (!req.body.user || !req.body.startDate || !req.body.endDate || req.body.createdAt) throw 'Invalid Body.'
-        UserModel.findById(req.body.user)
-            .then((user) => {
-                if (!user) throw `No such user ${req.body.organization}`
-                if (user.organization.toString() !== req.body.organization) throw 'No such user in organization'
-                if (new Date (req.body.startDate) > new Date (req.body.endDate)) throw 'Invalid Date.'
-                createNew(req, res, next, WFHModel)
-            })
-            .catch((error) => {
-                handleCatch(`${error}`, res, 401, next)
-            })
-    } catch (error) {
-        handleCatch(`${error}`, res, 401, next)
-    }
+    UserModel.findById(req.body.user)
+        .then((user) => {
+            if (!user) throw `No User Found`
+            if (req.body.status) throw 'Invalid Body.'
+            if (user.organization.toString() !== req.body.organization) throw 'No such user in organization'
+            if (new Date(req.body.startDate) > new Date(req.body.endDate)) throw 'Invalid Date.'
+            createNew(req, res, next, WFHModel)
+        })
+        .catch((error) => {
+            handleCatch(`${error}`, res, 401, next)
+        })
 }
 
 export const updateWFH = (req, res, next) => {
     try {
-        if (req.body.user || req.body.organization || req.body.status || req.body.createdAt) throw 'Invalid Body.'
-        if (req.body.startDate || req.body.endDate) {
-            WFHModel.findById(req.params.id)
-                .then((wfh) => {
-                    if (!wfh) throw `No such WFH ${req.params.id}`
-                    if (req.body.startDate && req.body.endDate && (new Date (req.body.startDate) > new Date (req.body.endDate))) throw 'Invalid Date.'
-                    if (!req.body.startDate && req.body.endDate && new Date (wfh.startDate) > new Date (req.body.endDate)) throw 'Invalid Date.'
-                    if(req.body.startDate && !req.body.endDate && new Date (req.body.startDate) > new Date (wfh.endDate)) throw 'Invalid Date.'
+        if (req.body.user || req.body.organization || req.body.status) throw 'Invalid Body.'
+        WFHModel.findById(req.params.id)
+            .then((wfh) => {
+                if (!wfh) throw 'No such wfh'
+                if (wfh.status !== "pending") throw `cannot update this WFH`
+                if (req.body.startDate || req.body.endDate) {
+                    if (!wfh) throw 'No such WFH'
+                    if (req.body.startDate && req.body.endDate && (new Date(req.body.startDate) > new Date(req.body.endDate))) throw 'Invalid Date.'
+                    if (!req.body.startDate && req.body.endDate && new Date(wfh.startDate) > new Date(req.body.endDate)) throw 'Invalid Date.'
+                    if (req.body.startDate && !req.body.endDate && new Date(req.body.startDate) > new Date(wfh.endDate)) throw 'Invalid Date.'
                     updateById(req, res, next, WFHModel, 'WFH')
-                })
-                .catch((error) => {
-                    handleCatch(`${error}`, res, 401, next)
-                })
-        }
-        else updateById(req, res, next, WFHModel, 'WFH')
+                }
+                else updateById(req, res, next, WFHModel, 'WFH')
+            })
+            .catch((error) => {
+                handleCatch(`${error}`, res, 401, next)
+            })
     } catch (error) {
         handleCatch(`${error}`, res, 401, next)
     }
@@ -47,5 +45,13 @@ export const getWFH = (req, res, next) => {
 }
 
 export const deleteWFH = (req, res, next) => {
-    deleteById(req.params.id, res, next, WFHModel, 'WFH')
+    WFHModel.findById(req.params.id)
+        .then((wfh) => {
+            if (!wfh) throw 'No such WFH'
+            if (wfh.status !== "pending") throw `cannot delete this WFH`
+            deleteById(req.params.id, res, next, WFHModel, 'WFH')
+        })
+        .catch((error) => {
+            handleCatch(`${error}`, res, 401, next)
+        })
 }
